Extract ModelCollectionViewController base class
Part of prep work for table and outline view support, there's a bunch of shared code for all potential types of
scrollable views of a model collection. This just pulls out the non-CollectionView-specific parts of
CollectionViewController into a superclass. Followup diff(s) will use this to create the OutlineViewController
